+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/handler/resetter/AbstractResetter.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.notify.impl.handler.resetter;
 
 import com.dwarfeng.notify.stack.handler.Resetter;
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
 
 /**
  * 重置器的抽象实现。
@@ -16,6 +17,44 @@ public abstract class AbstractResetter implements Resetter {
     public void init(Context context) {
         this.context = context;
     }
+
+    @Override
+    public void start() throws HandlerException {
+        try {
+            doStart();
+        } catch (HandlerException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new HandlerException(e);
+        }
+    }
+
+    /**
+     * 启动方法的具体实现。
+     *
+     * @throws Exception 启动过程中发生的任何异常。
+     * @see Resetter#start()
+     */
+    protected abstract void doStart() throws Exception;
+
+    @Override
+    public void stop() throws HandlerException {
+        try {
+            doStop();
+        } catch (HandlerException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new HandlerException(e);
+        }
+    }
+
+    /**
+     * 停止方法的具体实现。
+     *
+     * @throws Exception 停止过程中发生的任何异常。
+     * @see Resetter#stop()
+     */
+    protected abstract void doStop() throws Exception;
 
     @Override
     public String toString() {