+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/handler/resetter/NeverResetter.java
@@ -12,11 +12,11 @@ import org.springframework.stereotype.Component;
 public class NeverResetter extends AbstractResetter {
 
     @Override
-    public void start() {
+    protected void doStart() {
     }
 
     @Override
-    public void stop() {
+    protected void doStop() {
     }
 
     @Override